+++ bin/online_shop.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
 import * as cdk from '@aws-cdk/core';
-import { OnlineShopStack } from '../lib/online_shop-stack';
+import { OnlineShopStack } from '../lib/online-shop-stack';
 
 const app = new cdk.App();
 new OnlineShopStack(app, 'OnlineShopStack');
