+++ lib/online-shop-stack.ts
