+++ test/online_shop.test.ts
@@ -3,7 +3,7 @@
 // import * as OnlineShop from '../lib/online_shop-stack';
 
 // example test. To run these tests, uncomment this file along with the
-// example resource in lib/online_shop-stack.ts
+// example resource in lib/online-shop-stack.ts
 test('SQS Queue Created', () => {
 //   const app = new cdk.App();
 //     // WHEN
